refactor: splitted vector into single modules
