+++ src/algebra/vector/index.ts
@@ -1,0 +1,11 @@
+export * from './add';
+export * from './distance';
+export * from './divide';
+export * from './dot';
+export * from './equals';
+export * from './magnitude';
+export * from './max';
+export * from './min';
+export * from './normalized';
+export * from './scale';
+export * from './subtract';
