+++ src/algebra/vector/vector.types.ts
@@ -1,0 +1,2 @@
+export type vectorN = number[];
+export type vectorNr = Readonly<vectorN>;
